Create 2D_HLF_problem-verifying_solutions.py
This code verifies that the quantum circuit implemented in this respository actually solves the 2D Hidden Linear Function problem. It does this by checking that the z value that the circuit outputs actually defines the the hidden linear function ((2 z^{T} x) mod 4)). For a given value of some parameter n, the code checks that the circuit solves the problem for all 2^{n} possible values of the vector b (an n-bit string).